REFACTOR: Isolar callbacks e mover entidades de casos de teste
