+++ domain/models/test_case/test_case_execution_history.py
@@ -6,10 +6,11 @@ from domain.models.named_entity import Entity
 
 
 class TestCaseExecutionHistory(Entity):
-    last_epoch: int
+    epoch: int
     test_case_id: ObjectId
     created_at: datetime
     model_name: str
     model_id: str
     gpu_description: str
     cpu_description: str
+    results
